move set & verify consent to h3
