--- packages/backend/src/routes/__tests__/consent.test.ts
@@ -1,272 +1,0 @@
-import { beforeEach, describe, expect, it } from 'vitest';
-import { c15tInstance } from '~/core';
-import { memoryAdapter } from '~/pkgs/db-adapters';
-import { DoubleTieError, ERROR_CODES } from '~/pkgs/results';
-import type { ConsentPolicy, PolicyType } from '~/schema';
-import type { C15TContext } from '~/types';
-import { setConsent } from '../set-consent';
-
-describe('Consent Endpoints', () => {
-	let context: C15TContext;
-
-	beforeEach(async () => {
-		const instance = await c15tInstance({
-			baseURL: 'http://localhost:3000',
-			database: memoryAdapter({}),
-			trustedOrigins: ['http://localhost:3000'],
-			secret: 'test-secret',
-		});
-
-		const contextResult = await instance.$context;
-		if (!contextResult.isOk()) {
-			throw new Error('Failed to initialize context');
-		}
-		context = contextResult.value;
-	});
-
-	describe('setConsent', () => {
-		// Helper function for common consent data
-		const createConsentData = (type: PolicyType, overrides = {}) => ({
-			type,
-			domain: 'example.com',
-			preferences: {
-				marketing: true,
-				analytics: true,
-			},
-			...overrides,
-		});
-
-		// Helper function for common response checks
-		const expectValidConsentResponse = (
-			response: {
-				id: string;
-				givenAt: string;
-				subjectId: string;
-				domainId: string;
-				type: string;
-				status: string;
-				domain: string;
-				metadata?: unknown;
-			},
-			type: string,
-			extraChecks = {}
-		) => {
-			// biome-ignore lint/suspicious/noMisplacedAssertion: its okay
-			expect(response).toMatchObject({
-				type,
-				status: 'active',
-				domain: 'example.com',
-				metadata: undefined,
-				...extraChecks,
-			});
-			// biome-ignore lint/suspicious/noMisplacedAssertion: its okay
-			expect(typeof response.id).toBe('string');
-			// biome-ignore lint/suspicious/noMisplacedAssertion: its okay
-			expect(response.givenAt).toBeDefined();
-			// biome-ignore lint/suspicious/noMisplacedAssertion: its okay
-			expect(response.subjectId).toBeDefined();
-			// biome-ignore lint/suspicious/noMisplacedAssertion: its okay
-			expect(response.domainId).toBeDefined();
-		};
-
-		it('should set cookie banner consent successfully', async () => {
-			const response = await setConsent({
-				context,
-				params: undefined,
-				query: undefined,
-				body: createConsentData('cookie_banner'),
-			});
-			expectValidConsentResponse(response, 'cookie_banner');
-		});
-
-		it('should set privacy policy consent successfully with external subject', async () => {
-			await context.registry.createSubject({
-				externalId: 'test-subject',
-				isIdentified: true,
-				identityProvider: 'test',
-			});
-
-			const response = await setConsent({
-				context,
-				params: undefined,
-				query: undefined,
-				body: createConsentData('privacy_policy', {
-					externalSubjectId: 'test-subject',
-				}),
-			});
-
-			expectValidConsentResponse(response, 'privacy_policy', {
-				externalSubjectId: 'test-subject',
-			});
-		});
-
-		describe('Policy ID handling', () => {
-			let policy: ConsentPolicy;
-
-			beforeEach(async () => {
-				policy = await context.registry.createConsentPolicy({
-					type: 'privacy_policy',
-					name: 'Test Privacy Policy',
-					version: '1.0',
-					content: 'Test content',
-					contentHash: 'test-hash',
-					effectiveDate: new Date(),
-					updatedAt: new Date(),
-					isActive: true,
-				});
-			});
-
-			it('should set consent with explicit policy ID', async () => {
-				const response = await setConsent({
-					context,
-					params: undefined,
-					query: undefined,
-					body: createConsentData('privacy_policy', { policyId: policy.id }),
-				});
-
-				expectValidConsentResponse(response, 'privacy_policy');
-			});
-
-			it('should find latest policy when policy ID is not provided', async () => {
-				const response = await setConsent({
-					context,
-					params: undefined,
-					query: undefined,
-					body: createConsentData('privacy_policy'),
-				});
-
-				expectValidConsentResponse(response, 'privacy_policy');
-			});
-
-			it('should error with invalid policy ID', async () => {
-				await expect(
-					setConsent({
-						context,
-						params: undefined,
-						query: undefined,
-						body: createConsentData('privacy_policy', {
-							policyId: 'invalid-id',
-						}),
-					})
-				).rejects.toMatchObject({
-					name: 'DoubleTieError',
-					code: ERROR_CODES.NOT_FOUND,
-					status: 404,
-				});
-			});
-		});
-
-		describe('Subject mapping validation', () => {
-			it('should validate that subjectId and externalSubjectId map to the same subject', async () => {
-				// Create a subject with external ID
-				const subject = await context.registry.createSubject({
-					externalId: 'subject-mapping-1',
-					isIdentified: true,
-					identityProvider: 'test',
-				});
-
-				if (!subject) {
-					throw new Error('Failed to create test subject');
-				}
-
-				// Test with matching IDs
-				const response = await setConsent({
-					context,
-					params: undefined,
-					query: undefined,
-					body: createConsentData('privacy_policy', {
-						subjectId: subject.id,
-						externalSubjectId: 'subject-mapping-1',
-					}),
-				});
-
-				expectValidConsentResponse(response, 'privacy_policy', {
-					subjectId: subject.id,
-					externalSubjectId: 'subject-mapping-1',
-				});
-
-				// Create another subject with different external ID
-				const otherSubject = await context.registry.createSubject({
-					externalId: 'subject-mapping-2',
-					isIdentified: true,
-					identityProvider: 'test',
-				});
-
-				if (!otherSubject) {
-					throw new Error('Failed to create other test subject');
-				}
-
-				// Test with mismatched IDs
-				await expect(
-					setConsent({
-						context,
-						params: undefined,
-						query: undefined,
-						body: createConsentData('privacy_policy', {
-							subjectId: subject.id,
-							externalSubjectId: 'subject-mapping-2',
-						}),
-					})
-				).rejects.toMatchObject(
-					new DoubleTieError(
-						'The provided subjectId and externalSubjectId do not match the same subject. Please ensure both identifiers refer to the same subject.',
-						{
-							code: ERROR_CODES.CONFLICT,
-							status: 409,
-						}
-					)
-				);
-			});
-		});
-
-		describe('Error cases', () => {
-			it('should create anonymous subject when external subject is not found', async () => {
-				const response = await setConsent({
-					context,
-					params: undefined,
-					query: undefined,
-					body: createConsentData('privacy_policy', {
-						externalSubjectId: 'non-existent',
-					}),
-				});
-
-				expectValidConsentResponse(response, 'privacy_policy', {
-					externalSubjectId: 'non-existent',
-				});
-			});
-
-			it('should error if subject ID is not found', async () => {
-				await expect(
-					setConsent({
-						context,
-						params: undefined,
-						query: undefined,
-						body: createConsentData('privacy_policy', {
-							subjectId: 'non-existent',
-						}),
-					})
-				).rejects.toMatchObject({
-					name: 'DoubleTieError',
-					code: ERROR_CODES.NOT_FOUND,
-					status: 404,
-				});
-			});
-
-			it('should handle invalid consent data', async () => {
-				await expect(
-					setConsent({
-						context,
-						params: undefined,
-						query: undefined,
-						body: createConsentData('cookie_banner', {
-							preferences: {
-								marketing: 'invalid' as unknown as boolean,
-								analytics: 'invalid' as unknown as boolean,
-							},
-						}),
-					})
-				).rejects.toThrow('Invalid body parameters');
-			});
-		});
-	});
-});
